refactored with proper comments and logical placement of test-utils
