+++ src/debug/index.ts
@@ -1,2 +1,1 @@
 export * from './debugDynamo';
-export * from './debugDynamoTests';
